refactor: cartPage 안쓰는 컴포넌트 제거 및 폴더 구조 정리 - [양방향 import 제거] 초기 DB 상수 각각 useProduct, useCoupon으로 옮기기 - [코드 응집도 높히기] 재사용성 없는 컴포넌트는 한 파일에 합치기 - [도메인 분리] useCart에 Coupon 도메인 관련 useState를 useCoupon 도메인 분리 - cart 유틸 함수에 안쓰는 addToCart 제거
