+++ src/refactoring/hooks/useCoupon.ts
@@ -5,12 +5,33 @@ const addCouponToList = (coupons: Coupon[], newCoupon: Coupon): Coupon[] => {
   return [...coupons, newCoupon];
 };
 
+export const initialCoupons: Coupon[] = [
+  {
+    name: '5000원 할인 쿠폰',
+    code: 'AMOUNT5000',
+    discountType: 'amount',
+    discountValue: 5000,
+  },
+  {
+    name: '10% 할인 쿠폰',
+    code: 'PERCENT10',
+    discountType: 'percentage',
+    discountValue: 10,
+  },
+];
+
 export const useCoupons = (initialCoupons: Coupon[]) => {
   const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
+  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
   const addCoupon = (newCoupon: Coupon) => {
     setCoupons((prevCoupons) => addCouponToList(prevCoupons, newCoupon));
   };
-
-  return { coupons, addCoupon };
+  return {
+    coupons,
+    selectedCoupon,
+    setCoupons,
+    setSelectedCoupon,
+    addCoupon,
+  };
 };
