+++ src/refactoring/pages/carts/components/product/ProductCard.tsx
@@ -1,18 +1,21 @@
 import { CartItem, Product } from '../../../../../types';
 import { getRemainingStock, getMaxDiscount } from '../../../../models/cart';
 
-export interface ProductItemProps {
+export interface ProductCardProps {
   product: Product;
   cart: CartItem[];
   addToCart: (product: Product) => void;
 }
 
-const ProductItem = ({ product, cart, addToCart }: ProductItemProps) => {
+export const ProductCard = ({ product, cart, addToCart }: ProductCardProps) => {
   const remainingStock = getRemainingStock({ cart, product });
 
   const isSoldOut = remainingStock <= 0;
   return (
-    <div data-testid={`product-${product.id}`} className="bg-white p-3 rounded shadow">
+    <div
+      key={product.id}
+      data-testid={`product-${product.id}`}
+      className="bg-white p-3 rounded shadow">
       <div className="flex justify-between items-center mb-2">
         <span className="font-semibold">{product.name}</span>
         <span className="text-gray-600">{product.price.toLocaleString()}원</span>
@@ -47,8 +50,4 @@ const ProductItem = ({ product, cart, addToCart }: ProductItemProps) => {
       </button>
     </div>
   );
-};
-
-const ProductList = () => {
-  return <></>;
 };
